requirements.txt
Django==3.1
psycopg2-binary
BeautifulSoup4
google==2.0.3
spotipy
pandas
djangorestframework